registry/storage/genericmanifesthandler.go
package storage

import (
	"context"
	"fmt"

	"github.com/docker/distribution"
	dcontext "github.com/docker/distribution/context"
	"github.com/docker/distribution/manifest/generic"
	"github.com/opencontainers/go-digest"
)

type genericManifestHandler struct {
	blobStore  distribution.BlobStore
	repository distribution.Repository
	ctx        context.Context
}

var _ ManifestHandler = &genericManifestHandler{}

func (ms *genericManifestHandler) Unmarshal(ctx context.Context, dgst digest.Digest, content []byte) (distribution.Manifest, error) {
	dcontext.GetLogger(ms.ctx).Debug("(*genericManifestHandler).Unmarshal")

	var m generic.DeserializedManifest
	if err := m.UnmarshalJSON(content); err != nil {
		return nil, err
	}

	return &m, nil
}

func (ms *genericManifestHandler) Put(ctx context.Context, manifest distribution.Manifest, skipDependencyVerification bool) (digest.Digest, error) {
	dcontext.GetLogger(ms.ctx).Debug("(*genericManifestHandler).Put")

	m, ok := manifest.(*generic.DeserializedManifest)
	if !ok {
		return "", fmt.Errorf("non-generic manifest put to genericManifestHandler: %T", manifest)
	}

	if err := ms.verifyManifest(ms.ctx, *m, skipDependencyVerification); err != nil {
		return "", err
	}

	mt, payload, err := m.Payload()
	if err != nil {
		return "", err
	}

	revision, err := ms.blobStore.Put(ctx, mt, payload)
	if err != nil {
		dcontext.GetLogger(ctx).Errorf("error putting payload into blobstore: %v", err)
		return "", err
	}

	return revision.Digest, nil
}

// verifyManifest ensures that the manifest content is valid from the
// perspective of the registry. As a policy, the registry only tries to store
// valid content, leaving trust policies of that content up to consumers.
func (ms *genericManifestHandler) verifyManifest(ctx context.Context, mnfst generic.DeserializedManifest, skipDependencyVerification bool) error {
	var errs distribution.ErrManifestVerification

	if skipDependencyVerification {
		return nil
	}

	blobsService := ms.repository.Blobs(ctx)

	for _, descriptor := range mnfst.References() {
		_, err := blobsService.Stat(ctx, descriptor.Digest)

		if err != nil {
			if err != distribution.ErrBlobUnknown {
				errs = append(errs, err)
			}

			// On error here, we always append unknown blob errors.
			errs = append(errs, distribution.ErrManifestBlobUnknown{Digest: descriptor.Digest})
		}
	}

	if len(errs) != 0 {
		return errs
	}

	return nil
}
